[Avi/mux/demux] Improve timestamp accuracy for AVI, prevents aborted encoding due to invalid timestamps
git-svn-id: svn://svn.berlios.de/avidemux/branches/avidemux_2.6_branch_mean@8961 30592fff-cfd4-0310-bf0e-8235a3c7be26

+++ avidemux_core/ADM_coreDemuxer/include/ADM_Video.h
@@ -81,7 +81,7 @@ virtual 	uint8_t		open(const char *name)=0;
 virtual 	uint8_t		close(void)=0;
 virtual         uint64_t        frameToUs(uint32_t frame)
                                 {
-                                    float f=frame;
+                                    double f=frame;
                                     f*=_videostream.dwScale;
                                     f/=_videostream.dwRate;
                                     f*=1000000.;
